+++ src/qcommon/cm_polylib.h
@@ -42,7 +42,7 @@ typedef struct
 #define MAX_MAP_BOUNDS			65535
 
 // you can define on_epsilon in the makefile as tighter
-#ifndef	ON_EPSILON
+#if !defined(ON_EPSILON)
 #define	ON_EPSILON	0.1f
 #endif